Move routeStatic back to previous location
