+++ controllers/toplevel/index.js
@@ -14,7 +14,7 @@ const router = express.Router();
 const appData = require('../../modules/appData');
 const contentApi = require('../../services/content-api');
 const surveyService = require('../../services/surveys');
-const routeStatic = require('../routeStatic');
+const routeStatic = require('../utils/routeStatic');
 const getSecret = require('../../modules/get-secret');
 const analytics = require('../../modules/analytics');
 const { heroImages } = require('../../modules/images');
